Indexing ClueWeb09 Wikipedia documents.
git-svn-id: https://groke.otago.ac.nz/ant/ant@940 7489a7b4-4588-4892-8acd-774bacfab51c

+++ source/search_engine_result.h
@@ -169,7 +169,7 @@ public:
 		heapk->min_update(which);
 	else if (new_val > accumulator_pointers[0]->get_rsv())
 		{
-		include_set->unsafe_unsetbit(accumulator_pointers[0] - accumulator);
+		include_set->unsafe_unsetbit((long)(accumulator_pointers[0] - accumulator));
 		heapk->min_insert(which);
 		include_set->unsafe_setbit(index);
 		}
